Fix breaking changes in Thrift metadata behavior
This commit addresses 6 breaking changes identified in the metadata
refactoring that changed Thrift output behavior:

1. xdbc_column_size: Now uses Thrift-provided precision values via
   SetPrecisionScaleAndTypeName instead of synthesized values
2. xdbc_decimal_digits: Now uses Thrift-provided scale values via
   SetPrecisionScaleAndTypeName instead of synthesized values
3. xdbc_num_prec_radix: Explicitly set to null to match original
   behavior (was always null, refactor added synthesized values)
4. xdbc_sql_data_type: Now uses Thrift colType directly to match
   original behavior
5. xdbc_char_octet_length: Explicitly set to null to match original
   behavior (was always null in original)
6. xdbc_datetime_sub: Explicitly set to null to match original
   behavior (was always null in original)

Also fixed BufferLength type from byte? to int? with range validation
to prevent overflow when casting to sbyte for Int8Array.

The fix preserves the original Thrift flow:
- Call SetPrecisionScaleAndTypeName to get server-provided precision/scale
- Use MetadataFieldPopulator for structure and field synthesis
- Override specific fields with Thrift-provided values

This ensures byte-identical output for Thrift while keeping shared
abstractions flexible for SEA (which can use full synthesis).

+++ csharp/src/Drivers/Apache/Hive2/Metadata/ColumnMetadataRecord.cs
@@ -72,7 +72,7 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2.Metadata
         /// Transfer size of the data in bytes (BUFFER_LENGTH in statement-based metadata)
         /// Nullable for most types, only applicable for certain binary types
         /// </summary>
-        public byte? BufferLength { get; set; }
+        public int? BufferLength { get; set; }
 
         /// <summary>
         /// Number of fractional digits (DECIMAL_DIGITS in statement-based metadata)
